+++ package.json
@@ -18,7 +18,9 @@
     "autoprefixer": "10.4.15",
     "clsx": "^2.0.0",
     "firebase": "^10.5.2",
+    "firebase-admin": "^11.11.1",
     "formik": "^2.4.5",
+    "lodash": "^4.17.21",
     "mathjs": "^12.1.0",
     "next": "^14.0.0",
     "next-auth": "^5.0.0-beta.3",
