+++ src/lib/config/firebaseAdmin.config.ts
@@ -1,0 +1,16 @@
+import admin from "firebase-admin";
+
+const app = admin.initializeApp({
+  credential: admin.credential.cert(process.env.SERVICE_ACCOUNT || ""),
+});
+
+const auth = app.auth();
+const db = app.firestore();
+
+const FirebaseAdmin = {
+  app,
+  auth,
+  db,
+};
+
+export default FirebaseAdmin;
